chore: refactor project build structure (#7)
This change updates NEStalgic to include a modular build structure.

Each target is given its own CMakeLists.txt file and the root CMakeLists.txt is updated to reference these to using the `add_subdirectory`  command.

1. The package is renamed to `nestalgic` to reflect the name change.
2. The `src` directory is updated to create a static library target of all of the src files in the emulator. The include directories are added at a PRIVATE scope (as these files are not meant to be linked with external modules).
3. The `tests` directory is given its own CMakeLists.txt file to configure testing. The google test lib is fetched at version 1.16.0. An executable named `nestalgic_tests` is created from all of the test source files and linked with the GTest lib. The nestalgic_tests target is updated to include the header files declared in /include for testing.

Configuring the package in this way creates a more scalable build structure and aligns with CMake best practices.