feat: add CPU implementation (#9)
This change adds a baseline implementation of the NES CPU. It publicly declares a `Clock` and `Reset` method and adds support for internal registers and status flags. In addition, an internal `CPUTestHelper` class is added to an `internal` namespace to help with testing.

As a baseline implementation, support for the `LDA` operation with `Immediate` addressing is added.